requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
crispy-bootstrap5==0.6
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.6
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-next-prev==1.1.0
django-recaptcha==3.0.0
django-redis==5.2.0
django-tinymce==3.4.0
idna==3.3
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-decouple==3.6
python3-openid==3.2.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
wrapt==1.14.1
